Maximum number of parameters per TAB is set to 50.
Added array extension to split arrays at a fixed size.
Maximum number of parameters per category is 150 - otherwise the setup is messed up.
Added validation for this restriction.
Maximum numbers per TAB is therefore 50.
(This is because of some errors in C++-code which interprets generated INI-File.)

Parameter number 404 is reserved for "machine type" definition. Added validation for this.

Removed some finished TODOs.