DRY: remove duplication for roles and software options
Introduce a module BinaryCodedValues which handles the common stuff:
- transform value into a number
- transform a number into appropriate values

Roles and SoftwareOptions change into classes including this new module.
There was a name collision with number: number is an attribute in Parameter and in BinaryCodedValues.
Including Role or SoftwareOptions as modules in Parameter led to this collision - changing the type resolves this issue.

Change usings of Role and SoftwareOptions.